test(e2e): cover messaging endpoint helpers
